Add files via upload
The template is written for both automatic and instant mailing (for justclick.ru).
To transfer a design from the file to justclick, just open it with a notepad, copy and paste in the "Source" mode.

Links to social networks are specific, you can replace them with the ones you need. Links to projects are specific. Can be replaced as well.

When an image is published in the text, upload a picture, click "advanced" tab, copy and paste in the very bottom line: max-width: 100%;